Render link when JSON key is `href`
This commit adds a test for JSON _key_ matching the word `href` where
there was already testing for a JSON _value_ matching a URL scheme.

As a heuristic, jsonview already renders links (`<a href … > … </a>`) when
JSON value begins with a known URL scheme (http, https, or file). However, this
heuristic misses relative URLs (In my case, specifically URLs with paths but
no authority part). A JSON _key_ of `href` may be a good (additional) heuristic
for when the corresponding _value_ is a URI.