Fix an issue where JSONView would not work on first load, and would require disable/reenable to work.
Basically, the method of broadcasting an initialization message to all frame scripts was prone to problems if a frame script loaded later, or if the frame script hadn't loaded by the time the message was sent. This simplifies things by having the frame script load itself rather than waiting for a message.