Do not read content type from redirect responses
The content type header is irrelevant for redirect (status 300 to 400)
responses, as they do not say anything about the content type of the
actual resource they redirect to.

This commit makes the extension silently ignore redirect responses. The
resource at the end of the redirect is still handled properly.

No special consideration is taken for status code 300 which will not
make the browser create a new request, as the response body for these
responses should already be made for display to the user.

Fixes #172